♻️ movie/[id]/page.tsx 분리 (406 → 58줄, 200줄 룰 준수)
페이지 라우트가 SEO 메타·JSON-LD 생성 로직 다수를 모두 담고 있어 406줄
스멜이었던 것을 책임별로 4개 파일로 분리.

- \`data.ts\` (40줄) — \`getMovieDetail\`, \`getReviews\`, \`getScore\` +
  \`hasValidScore\` 가드
- \`metadata.ts\` (143줄) — \`generateMovieMetadata\` (Next \`Metadata\` 객체 생성,
  키워드·description·OpenGraph·Twitter·other meta tags)
- \`json-ld.ts\` (139줄) — \`buildMovieJsonLd\` (Movie schema · Review ·
  AggregateRating · WatchAction), \`buildBreadcrumbJsonLd\` (BreadcrumbList)
- \`page.tsx\` (58줄) — 데이터 fetch + 두 JSON-LD 빌더 호출 + JSX 렌더링만

동작 변화 없음. 중복 \`hasValidScore\`, 중복 \`https://drunkenmovie.shop\`
리터럴 (\`BASE\` 상수) 정리로 코드량 26줄 감소.
